Prevent invisible pencil marks from being modified
We modify the page so that when user requests toggling a pencil mark on
a field where there's already a value provided, this is a no-op. Such
requests are mostly mistakes and we don't want to change the pencil
marks because if the number is deleted, the marks are brought back and
could be misleading if we allow modifying them by accident without any
feedback.