Skip all analysis when any value is incorrect
If a user makes a mistake and puts a value which is obviously wrong, we
don't want to mess up with pencil marks or automatically fill any
values. To make the code of the autofill functions simpler, we put
a guard against incorrect cells before calling any of them so that
individual autofill routines can be simpler by assuming that all values
are valid.