Switch from CSS to SCSS
It's annoying that we have to say `className="btn btn-outline-bark"` for
every button in the JS code to use Bootstrap button styles. Styling
should be part of the stylesheet, not the JS code.

To fix that, we switch from CSS to SCSS so that our stylesheet can say
that all our buttons inherit their style from Bootstrap CSS classes.