Reactor structure of the panel on the right
Using a mix of a grid and gridless layout makes it difficult to tell
that action buttons should span the whole width of the panel. We tried
doing that in d7c7b6f4220eeb but this didn't really work and the action
buttons were longer than needed. To fix that, we refactor the structure
of the panel so that all elements are placed in a single CSS grid.