Make it possible to select multiple cells
Adding pencil marks by hand is much more convenient when one can select
multiple cells and add the same marks to all of them. To make this
possible, we modify the game so that when a cell is clicked with Control
pressed, it is added or removed form the selection rather than replacing
it.