Fix loading static assets in the deployed page
The page, when deployed, is available as

    https://marcinsulikowski.github.io/sudoku-assistant-page/

This breaks loading assets like scripts and CSS stylesheets because the
page requests them as "static/xxxx" which results in requests like this:

    https://marcinsulikowski.github.io/static/xxx

To fix this, we need to tell React to create files suitable for serving
from path `sudoku-assistant-page/`. We achieve this by adding
`PUBLIC_URL` environment variable to the `npm run build` step.